# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Public API documentation served through QRUPI API Gateway.",
        "title": "QRUPI V2 API",
        "contact": {},
        "version": "2.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/users": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "User API"
                ],
                "summary": "Get all User",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Sort by field",
                        "name": "sortBy",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "asc",
                            "desc"
                        ],
                        "type": "string",
                        "description": "Sort order (asc or desc)",
                        "name": "sortOrder",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of results per page",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {}
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "User API"
                ],
                "summary": "Create User",
                "parameters": [
                    {
                        "type": "string",
                        "name": "avatar_url",
                        "in": "formData"
                    },
                    {
                        "maxLength": 100,
                        "type": "string",
                        "name": "context_code",
                        "in": "formData"
                    },
                    {
                        "maxLength": 50,
                        "type": "string",
                        "name": "context_type",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "name": "email",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "name": "institution_id",
                        "in": "formData"
                    },
                    {
                        "enum": [
                            "active",
                            "inactive"
                        ],
                        "type": "string",
                        "name": "is_active",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "maxLength": 255,
                        "minLength": 2,
                        "type": "string",
                        "name": "name",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "minLength": 6,
                        "type": "string",
                        "name": "password",
                        "in": "formData"
                    },
                    {
                        "maxLength": 20,
                        "type": "string",
                        "name": "phone",
                        "in": "formData"
                    },
                    {
                        "maxLength": 8,
                        "minLength": 4,
                        "type": "string",
                        "name": "pin",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "name": "role_id",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "enum": [
                            "student",
                            "teacher",
                            "admin",
                            "staff"
                        ],
                        "type": "string",
                        "name": "type",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "Upload File (optional)",
                        "name": "file",
                        "in": "formData"
                    }
                ],
                "responses": {}
            }
        },
        "/users/import-excel": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "User API"
                ],
                "summary": "Import Users by Excel",
                "parameters": [
                    {
                        "type": "file",
                        "description": "Excel file (.xlsx)",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
...
        "/users/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "User API"
                ],
                "summary": "Get by id User",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "User API"
                ],
                "summary": "Update User",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "avatar_url",
                        "in": "formData"
                    },
                    {
                        "maxLength": 100,
                        "type": "string",
                        "name": "context_code",
                        "in": "formData"
                    },
                    {
                        "maxLength": 50,
                        "type": "string",
                        "name": "context_type",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "name": "email",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "name": "institution_id",
                        "in": "formData"
                    },
                    {
                        "enum": [
                            "active",
                            "inactive"
                        ],
                        "type": "string",
                        "name": "is_active",
                        "in": "formData"
                    },
                    {
                        "maxLength": 255,
                        "minLength": 2,
                        "type": "string",
                        "name": "name",
                        "in": "formData"
                    },
                    {
                        "minLength": 6,
                        "type": "string",
                        "name": "password",
                        "in": "formData"
                    },
                    {
                        "maxLength": 20,
                        "type": "string",
                        "name": "phone",
                        "in": "formData"
                    },
                    {
                        "maxLength": 8,
                        "minLength": 4,
                        "type": "string",
                        "name": "pin",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "name": "role_id",
                        "in": "formData"
                    },
                    {
                        "enum": [
                            "student",
                            "teacher",
                            "admin",
                            "staff"
                        ],
                        "type": "string",
                        "name": "type",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "Upload File (optional)",
                        "name": "file",
                        "in": "formData"
                    }
                ],
                "responses": {}
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "User API"
                ],
                "summary": "Delete User permanently",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/users/{id}/archived": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "User API"
                ],
                "summary": "Archive (soft delete) User",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/users/{id}/barcode": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "image/png"
                ],
                "tags": [
                    "User API"
                ],
                "summary": "Download student barcode",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Student user ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "file"
                        }
                    }
                }
            }
        },
...
    },
...
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}